+++ cmd/helm/downloader/chart_downloader.go
@@ -24,7 +24,6 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"path"
 	"path/filepath"
 	"strings"
 
@@ -174,24 +173,6 @@ func (c *ChartDownloader) ResolveChartVersion(ref, version string) (*url.URL, er
 		return u, fmt.Errorf("chart %q has no downloadable URLs", ref)
 	}
 	return url.Parse(cv.URLs[0])
-}
-
-// urlJoin joins a base URL to one or more path components.
-//
-// It's like filepath.Join for URLs. If the baseURL is pathish, this will still
-// perform a join.
-//
-// If the URL is unparsable, this returns an error.
-func urlJoin(baseURL string, paths ...string) (string, error) {
-	u, err := url.Parse(baseURL)
-	if err != nil {
-		return "", err
-	}
-	// We want path instead of filepath because path always uses /.
-	all := []string{u.Path}
-	all = append(all, paths...)
-	u.Path = path.Join(all...)
-	return u.String(), nil
 }
 
 func findRepoEntry(name string, repos []*repo.Entry) (*repo.Entry, error) {
