+++ pkg/repo/index.go
@@ -22,7 +22,9 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
+	"net/url"
 	"os"
+	"path"
 	"path/filepath"
 	"sort"
 	"strings"
@@ -96,8 +98,12 @@ func NewIndexFile() *IndexFile {
 func (i IndexFile) Add(md *chart.Metadata, filename, baseURL, digest string) {
 	u := filename
 	if baseURL != "" {
+		var err error
 		_, file := filepath.Split(filename)
-		u = strings.TrimSuffix(baseURL, "/") + "/" + file
+		u, err = urlJoin(baseURL, file)
+		if err != nil {
+			u = filepath.Join(baseURL, file)
+		}
 	}
 	cr := &ChartVersion{
 		URLs:     []string{u},
@@ -294,4 +300,22 @@ func LoadIndexFile(path string) (*IndexFile, error) {
 		return nil, err
 	}
 	return LoadIndex(b)
+}
+
+// urlJoin joins a base URL to one or more path components.
+//
+// It's like filepath.Join for URLs. If the baseURL is pathish, this will still
+// perform a join.
+//
+// If the URL is unparsable, this returns an error.
+func urlJoin(baseURL string, paths ...string) (string, error) {
+	u, err := url.Parse(baseURL)
+	if err != nil {
+		return "", err
+	}
+	// We want path instead of filepath because path always uses /.
+	all := []string{u.Path}
+	all = append(all, paths...)
+	u.Path = path.Join(all...)
+	return u.String(), nil
 }
